Split OffloaderController: extract rebind
Extracts the offloader-side endpoint-rebind cluster into a
new sibling module ``controllers/remote_build/rebind.py``:
``probe_pairing_endpoint``, ``commit_endpoint_rebind``,
``maybe_schedule_rebind_probe``, ``probe_and_rebind_endpoint``,
plus the module-private ``_respawn_peer_link_at_new_endpoint``,
``_clear_cooldown_on_unexpected_exit``, and
``_fire_offloader_pair_endpoint_rebound`` helpers. Bodies take
``OffloaderController`` as the first arg; the controller keeps
thin bound-method delegates for the four entry points that
either tests or cross-module callers reach into:
- ``_probe_pairing_endpoint`` (tests + ``edit_pairing_endpoint``)
- ``_commit_endpoint_rebind`` (tests + ``edit_pairing_endpoint``)
- ``_maybe_schedule_rebind_probe`` (``discovery.upsert_host``
  routes through the delegate, tests instance-call)
- ``_probe_and_rebind_endpoint`` (tests instance-call)

The strictly-internal helpers stay module-private (no tests
or cross-module callers reach them) and are called as module
functions from within ``rebind.py``.

Imports moved out of ``offloader.py``: ``Iterator`` /
``contextmanager`` (only ``_clear_cooldown_on_unexpected_exit``
used them), ``normalize_hostname``,
``OffloaderPairEndpointReboundData``, the
``_REBIND_PROBE_COOLDOWN_SECONDS`` constant. ``RebindProbeOutcome``
/ ``RebindProbeResult`` / ``peer_link_preview_pair`` /
``PeerLinkClientError`` stay imported in ``offloader.py``
because ``edit_pairing_endpoint`` and ``preview_pair`` still
use them directly.

Test repoint: 4 ``monkeypatch.setattr(rb, "peer_link_preview_pair", ...)``
sites move to ``rb_rebind`` (the new rebind module alias).
All four patches exercise paths that route through
``_probe_pairing_endpoint``, whose body now lives in
``rebind.py``; the standalone ``preview_pair`` WS command
path stays on ``offloader.py`` and keeps its own
``peer_link_preview_pair`` import.

Offloader: 1597 -> 1429 lines (-168). 214 controller-test
+ 676 remote-build tests pass.
